Fix error when canceling Appointments that don't require a response
When declining meetings that don't require a response, Nothing is returned from the Respond method. As no response is possible, the cancelMsg is skipped for those meetings.

Also, changed the message prompt to include the time of the date range, so it becomes apparent that while the endDate is included, no events pass midnight are cancelled.